Struct mcmu::buff::Buff

pub struct Buff<T> { /* fields omitted */ }

Buffer data structure for parsing

Implementations

impl<T> Buff<T> where
T: Display,
T: Eq,
T: Clone,

pub fn new(data: Vec<T>) -> Self

Create a new buffer

pub fn save(&mut self)

Push the current position to the stack

pub fn restore(&mut self)

Pop the last position pushed to the stack and set the current position to it

pub fn update_save(&mut self)

Update the top of the stack to be the current position

pub fn top(&self) -> Option<T>

Get the first element of the buffer and returns None if the buffer is empty

pub fn is_empty(&self) -> bool

Check if there are still elements to read in the buffer

pub fn pop(&mut self) -> Option<()>

Drop the first element of the buffer and returns None if the buffer is empty

pub fn next(&mut self) -> Option<T>

Get the first element of the buffer and drops it. Returns None if the buffer is empty

pub fn expect(&mut self, x: T) -> Option<()>

Compare a given element with the first element of the buffer and drops it if they are equals. Returns None if the buffer is empty

pub fn expect_cond<P>(&mut self, cond: P) -> Option<T> where
P: FnOnce(&T) -> bool,

Expect the current element of the buffer to satisfies a condition and returns it.

pub fn expect_convert<U>(&mut self, conv: fn(_: T) -> Option<U>) -> Option<U>

Expect the current element of the buffer to be convertible. The conversion function is expected to return None if the element is not convertible

pub fn convert_while<U>(
&mut self,
pre: fn(_: &T) -> bool,
conv: fn(_: T) -> U
) -> Option<Vec<U>>

Expect a non-empty sequence of elements satisfying a given condition. Elements are converted according to a given conversion function.

pub fn convert_list<U>(&mut self, conv: fn(_: T) -> Option<U>) -> Option<Vec<U>>

Expect a non-empty sequence of elements satisfying a given condition. The condition is given as a conversion function returning None if the element violates the condition

pub fn expect_list<U>(
&mut self,
parse: fn(_: &mut Buff<T>) -> Option<U>
) -> Option<Vec<U>>

Parse a non-empty list of elements according to a parsing function

pub fn expect_end(&self) -> Option<()>

pub fn expect_one_of(&mut self, alt: Vec<T>) -> Option<()>

impl Buff<char>

pub fn trim(&mut self)

pub fn expect_u32(&mut self) -> Option<u32>

pub fn expect_alpha(&mut self) -> Option<char>

pub fn expect_digit(&mut self) -> Option<char>

pub fn expect_symb(&mut self) -> Option<String>

pub fn expect_blank(&mut self) -> Option<()>

pub fn expect_token(&mut self, tok: String) -> Option<()>